+++ crates/vm/src/system/cuda/memory.rs
@@ -234,8 +234,12 @@ impl MemoryInventoryGPU {
 
 impl Drop for PersistentMemoryInventoryGPU {
     fn drop(&mut self) {
-        // Drop merkle subtrees first so their individual streams synchronize before dropping the
+        // Force synchronize all streams in merkle tree before dropping the
         // initial memory buffers. This prevents buffers from dropping before build_async completes.
+        for s in &self.merkle_tree.subtrees {
+            s.stream.synchronize().unwrap();
+        }
+        self.merkle_tree.stream.synchronize().unwrap();
         self.merkle_tree.drop_subtrees();
         self.initial_memory.clear();
     }